requirements1.txt
# Backend
fastapi
uvicorn[standard]
scikit-learn
pandas
nltk
faiss-cpu
transformers
torch
python-multipart
imbalanced-learn
openai

# Dashboard
streamlit
matplotlib
seaborn

# Data & HF Integration
datasets